Update CropDialog.kt
Refactored a bit for code size. Can only add a unique our of the day once to the watering times. Time will be returned to list when undo is pressed, but is returned to bottom of list. Also getting strange behavior that definitely needs fixed. Continuously autoselecting items from watering hour dropdown.